Refactored testing setup: - renamed phpunit.xml to phpunit.xml.dist (supported by phpunit) - added phpunit.xml to the .gitignore - added commented out redis configuration - Redis based tests are skipped when configuration not present

+++ tests/Queue/RedisQueueIntegrationTest.php
@@ -21,11 +21,18 @@ class RedisQueueIntegrationTest extends PHPUnit_Framework_TestCase
     public function setUp()
     {
         parent::setUp();
+        
+        $host = getenv('REDIS_HOST');
+        $port = getenv('REDIS_PORT');
+        if (!$host) {
+            $this->markTestSkipped('Set environment variable REDIS_HOST & REDIS_PORT to enable ' . __CLASS__);
+            return;
+        }
         $this->redis = new Database([
             'cluster' => false,
             'default' => [
-                'host' => '127.0.0.1',
-                'port' => 6379,
+                'host' => $host,
+                'port' => $port,
                 'database' => 5,
             ],
         ]);
@@ -39,7 +46,10 @@ class RedisQueueIntegrationTest extends PHPUnit_Framework_TestCase
     {
         parent::tearDown();
         m::close();
-        $this->redis->connection()->flushdb();
+        if ($this->redis) {
+            $this->redis->connection()->flushdb();            
+        }
+
     }
 
     public function testExpiredJobsArePopped()
